refactory - refatoração do metodo de criação de players para um facotry pattern, e o metodo de leitura de arquivos para um strategy pattern
