+++ OTServer.Domain/Factory/PlayerType.cs
@@ -1,0 +1,11 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace OTServer.Domain.Factory
+{
+    public enum PlayerType
+    {
+        basicPlayer = 1
+    }
+}
