+++ OTServer.Domain/Strategy/IReader.cs
@@ -1,0 +1,11 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace OTServer.Domain.Strategy
+{
+    public interface IReader<T>
+    {
+        List<T> ReaderFiles(string directory);
+    }
+}
