+++ OTServer.UI.MVC/Controllers/DownloadController.cs
@@ -11,7 +11,7 @@ namespace OTServer.UI.MVC.Controllers
     {
         public IActionResult Index()
         {
-            string file = @"~/OLD TIMES 1.0 CLIENTE.rar";
+            string file = @"~/Ilusion Install v1.exe";
             string contentType = "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet";
             return File(file, contentType, Path.GetFileName(file));
         }
