+++ OTServer.UI.MVC/Properties/PublishProfiles/FolderProfile.pubxml
@@ -13,7 +13,7 @@ by editing this MSBuild file. In order to learn more about this please visit htt
     <LaunchSiteAfterPublish>True</LaunchSiteAfterPublish>
     <ExcludeApp_Data>False</ExcludeApp_Data>
     <ProjectGuid>5b637a5e-7b17-4780-9812-0d356a057390</ProjectGuid>
-    <publishUrl>D:\Area de trabalho\CSharp\server\setup</publishUrl>
+    <publishUrl>D:\Area de trabalho\CSharp\server\deploy</publishUrl>
     <DeleteExistingFiles>True</DeleteExistingFiles>
     <TargetFramework>netcoreapp3.1</TargetFramework>
     <RuntimeIdentifier>win-x64</RuntimeIdentifier>
