+++ OTServer.UI.MVC/Properties/PublishProfiles/FolderProfile.pubxml.user
@@ -6,6 +6,6 @@ by editing this MSBuild file. In order to learn more about this please visit htt
 <Project ToolsVersion="4.0" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
   <PropertyGroup>
     <TimeStampOfAssociatedLegacyPublishXmlFile />
-    <_PublishTargetUrl>D:\Area de trabalho\CSharp\server\setup</_PublishTargetUrl>
+    <_PublishTargetUrl>D:\Area de trabalho\CSharp\server\deploy</_PublishTargetUrl>
   </PropertyGroup>
 </Project>
