--- OTServer.UI.MVC/Views/Account/CriarPersonagem.cshtml
@@ -1,46 +1,0 @@
-﻿@model OTServer.UI.MVC.Models.DTOCriarPersonagem
-
-<div class="modal-dialog modal-dialog-centered" role="document">
-    <div class="modal-content">
-        <div class="modal-body">
-            <form asp-action="CriarPersonagem">
-                <h4 class="text-center text-info">Criar Personagem</h4>
-                <div asp-validation-summary="ModelOnly" class="text-danger"></div>
-                <div class="form-group">
-                    <label asp-for="Nome" class="control-label"></label>
-                    <input asp-for="Nome" class="form-control" pattern="^[A-Za-z]{5,20}" required />
-                    <span asp-validation-for="Nome" class="text-danger"></span>
-                </div>
-                <div class="form-group">
-                    <div class="custom-control custom-radio">
-                        <input type="radio" id="customRadio1" name="Sexo" value="0" asp-for="Sexo" class="custom-control-input" checked="">
-                        <label class="custom-control-label" for="customRadio1">Female</label>
-                    </div>
-                    <div class="custom-control custom-radio">
-                        <input type="radio" id="customRadio2" name="Sexo" value="1" asp-for="Sexo" class="custom-control-input">
-                        <label class="custom-control-label" for="customRadio2">Male</label>
-                    </div>
-                </div>
-                <div class="form-group">
-                    <label asp-for="Voc" class="control-label"></label>
-                    <select class="custom-select" asp-for="Voc" required>
-                        <option value="">Selecione sua vocação</option>
-                        <option value="1">Sorcerer</option>
-                        <option value="2">Druid</option>
-                        <option value="3">Archer</option>
-                        <option value="3">Knight</option>
-                    </select>
-                </div>
-                <div class="form-group">
-                    <input type="submit" value="Criar Personagem" class="btn btn-primary btn-block" />
-                </div>
-            </form>
-        </div>
-        <div class="modal-footer">
-            <button type="button" class="btn btn-secondary" data-dismiss="modal">Close</button>
-        </div>
-    </div>
-</div>
-@section Scripts {
-    @{await Html.RenderPartialAsync("_ValidationScriptsPartial");}
-}
