+++ OTServer.UI.MVC/appsettings.json
@@ -7,17 +7,17 @@
     }
   },
   "ConnectionStrings": {
-    //"CaminhoServer": "C:\\Users\\Administrator\\Desktop\\OldTimes\\data",
-    //"CaminhoAccounts": "C:\\Users\\Administrator\\Desktop\\OldTimes\\data\\accounts",
-    //"CaminhoPlayers": "C:\\Users\\Administrator\\Desktop\\OldTimes\\data\\players",
-    //"CaminhoGuilds": "C:\\Users\\Administrator\\Desktop\\OldTimes\\data\\guilds.xml",
-    //"CaminhoOnlineList": "C:\\Users\\Administrator\\Desktop\\OldTimes\\data\\logs\\online.php",
+    //"CaminhoServer": "C:\\Users\\Administrator\\Desktop\\Ilusion 3-30-2020\\data",
+    //"CaminhoAccounts": "C:\\Users\\Administrator\\Desktop\\Ilusion 3-30-2020\\data\\accounts",
+    //"CaminhoPlayers": "C:\\Users\\Administrator\\Desktop\\Ilusion 3-30-20200\\data\\players",
+    //"CaminhoGuilds": "C:\\Users\\Administrator\\Desktop\\Ilusion 3-30-2020\\data\\guilds.xml",
+    //"CaminhoOnlineList": "C:\\Users\\Administrator\\Desktop\\Ilusion 3-30-2020\\data\\logs\\online.php",
     "CaminhoServer": "D:\\WorldStyle\\data",
     "CaminhoAccounts": "D:\\WorldStyle\\data\\accounts",
     "CaminhoPlayers": "D:\\WorldStyle\\data\\players",
     "CaminhoGuilds": "D:\\WorldStyle\\data\\guilds.xml",
     "CaminhoOnlineList": "D:\\WorldStyle\\data\\logs\\online.php",
-    "NomeDoServer": "New Old Times RPG"
+    "NomeDoServer": "Ilusion Pex"
   },
